requirements.txt
langchain-groq
streamlit
python-dotenv
fastapi
uvicorn
pydantic